requirements.txt
google-adk
fastapi
uvicorn
requests
python-dotenv
pydantic
